+++ src/containers/DaedalusSimpleContainer.js
@@ -8,16 +8,15 @@ import Alert from '../components/Alert';
 import { YoroiCallback } from '../API/yoroi';
 
 import { DesktopOnly, MobileOnly } from '../components/layout/Breakpoints';
-import { getPools, listPools } from '../API/api';
-import type { ApiPoolsResponse, Pool, SearchParams } from '../API/api';
-import AdapoolsDesktopTable from '../components/tables/AdapoolsDesktopTable';
-import AdapoolsMobileTable from '../components/tables/AdapoolsMobileTable';
-import SortSelect from '../components/SortSelect';
+import { getPoolsByDaedalusSimple } from '../API/api';
+import type { ApiPoolsDaedalusSimpleResponse, PoolDaedalusSimple, SearchParams } from '../API/api';
 import type { QueryState } from '../utils/types';
 
 import Modal from '../components/common/Modal';
 import SaturatedPoolAlert from '../components/SaturatedPoolAlert';
 import ProviderSelect from '../components/ProviderSelect';
+import DaedalusSimpleDesktopTable from '../components/tables/DaedalusSimpleDesktopTable';
+import DaedalusSimpleMobileTable from '../components/tables/DaedalusSimpleMobileTable';
 
 // k = 500
 const SATURATION_LIMIT = 63600000000000;
@@ -35,78 +34,47 @@ const Header = styled.div`
   }
 `;
 
-// const ColorButton = styled.button`
-//   border: none;
-//   background: none;
-//   color: #2B2C32;
-//   font-size: 14px;
-//   line-height: 22px;
-//   text-decoration: underline;
-//   margin-left: auto;
-//   cursor: pointer;
-//   @media (max-width: 1125px){
-//     margin-top: 30px;
-//   }
-// `;
-
 export type UrlParams = {|
-    chromeId: ?string,
-    mozId: ?string,
-    source: ?string,
-    selectedPoolIds: ?Array<string>,
-    lang: ?string,
-    totalAda: ?number,
+  chromeId: ?string,
+  mozId: ?string,
+  source: ?string,
+  selectedPoolIds: ?Array<string>,
+  lang: ?string,
+  totalAda: ?number,
 |};
 
-export type AdapoolsTableProps = {|
+export type DaedalusSimpleTableProps = {|
   urlParams: UrlParams,
 |};
 
 export type DelegationProps = {|
-    stakepoolName: string,
-    stakepoolTotalStake: string,
-    isAlreadySaturated: boolean,
-    id: string,
-|}
+  stakepoolName: string,
+  stakepoolTotalStake: string,
+  isAlreadySaturated: boolean,
+  id: string,
+|};
 
-function AdapoolsTable(props: AdapoolsTableProps): Node {
-  const [rowData, setRowData] = React.useState<?Array<Pool>>(null);
+function DaedalusSimpleContainer(props: DaedalusSimpleTableProps): Node {
+  const [rowData, setRowData] = React.useState<?Array<PoolDaedalusSimple>>(null);
   const [status, setStatus] = React.useState<QueryState>('idle');
   const [filterOptions, setFilterOptions] = React.useState<SearchParams>({
     search: '',
-    sort: 'score',
   });
   const [openModal, setOpenModal] = React.useState<boolean>(false);
   const [confirmDelegationModal, setConfirmDelegationModal] = React.useState<boolean>(false);
   const [delegationModalData, setDelegationModalData] = React.useState<Object>({});
 
-  const toPoolArray: ?{| [string]: Pool |} => Array<Pool> = (pools) => {
+  const toPoolArray: (?{| [string]: PoolDaedalusSimple |}) => Array<PoolDaedalusSimple> = (
+    pools,
+  ) => {
     if (pools == null) return [];
-    return Object.keys(pools).map(key => pools[key]);
+    return Object.keys(pools).map((key) => pools[key]);
   };
 
   useEffect(() => {
     setStatus('pending');
-    listPools()
-      .then((poolsData: ApiPoolsResponse) => {
-        setStatus('resolved');
-        setRowData(toPoolArray(poolsData.pools))
-      })
-      .catch((err) => {
-        setStatus('rejected');
-        console.error(err);
-      });
-  }, []);
-
-  const filterSelect = (value) => {
-    const newSearch = {
-      ...filterOptions,
-      sort: value,
-    };
-    setFilterOptions(newSearch);
-    setStatus('pending');
-    getPools(newSearch)
-      .then((poolsData: ApiPoolsResponse) => {
+    getPoolsByDaedalusSimple()
+      .then((poolsData: ApiPoolsDaedalusSimpleResponse) => {
         setStatus('resolved');
         setRowData(toPoolArray(poolsData.pools));
       })
@@ -114,7 +82,8 @@ function AdapoolsTable(props: AdapoolsTableProps): Node {
         setStatus('rejected');
         console.error(err);
       });
-  };
+  }, []);
+
   const filterSearch = (value) => {
     const newSearch = {
       ...filterOptions,
@@ -122,8 +91,8 @@ function AdapoolsTable(props: AdapoolsTableProps): Node {
     };
     setFilterOptions(newSearch);
     setStatus('pending');
-    getPools(newSearch)
-      .then((poolsData: ApiPoolsResponse) => {
+    getPoolsByDaedalusSimple()
+      .then((poolsData: ApiPoolsDaedalusSimpleResponse) => {
         setStatus('resolved');
         setRowData(toPoolArray(poolsData.pools));
       })
@@ -136,32 +105,33 @@ function AdapoolsTable(props: AdapoolsTableProps): Node {
   const confirmedDelegateFunction = (id: string): void => {
     const { urlParams } = props;
 
-    YoroiCallback(([id]), {
+    YoroiCallback([id], {
       source: urlParams.source,
       chromeId: urlParams.chromeId,
       mozId: urlParams.mozId,
     });
-  }
+  };
 
   const delegateFunction = (delegation: DelegationProps, totalAda: ?number): void => {
     if (delegation == null) return;
     const lovelaceDelegation = totalAda == null ? 0 : totalAda * 1000000;
 
     if (Number(delegation.stakepoolTotalStake) + lovelaceDelegation >= SATURATION_LIMIT) {
-      setDelegationModalData({ ...delegation, totalAda })
-      setConfirmDelegationModal(true)
-      setOpenModal(true)
+      setDelegationModalData({ ...delegation, totalAda });
+      setConfirmDelegationModal(true);
+      setOpenModal(true);
     } else {
-      confirmedDelegateFunction(delegation.id)
+      confirmedDelegateFunction(delegation.id);
     }
   };
 
-  const alertText = 'The new saturation point for Stakepools will be 63.6 million ADA from December 6th. If the "Pool Size" parameter of your Stakepool is over this limit, delegate to a new stakepool to avoid less than expected rewards';
+  const alertText =
+    'The new saturation point for Stakepools will be 63.6 million ADA from December 6th. If the "Pool Size" parameter of your Stakepool is over this limit, delegate to a new stakepool to avoid less than expected rewards';
 
   function filterPools(
-    pools: ?Array<Pool>,
+    pools: ?Array<PoolDaedalusSimple>,
     totalAda: ?number,
-  ): ?Array<Pool> {
+  ): ?Array<PoolDaedalusSimple> {
     if (pools == null) return pools;
 
     // don't filter out saturated pools if the user explicitly searches
@@ -173,25 +143,25 @@ function AdapoolsTable(props: AdapoolsTableProps): Node {
 
     if (lovelaceDelegation > SATURATION_LIMIT) return pools;
 
-    return pools.filter(item => (
-      Number(item.total_stake) + lovelaceDelegation < SATURATION_LIMIT
-    ));
+    // const arr = pools.filter(
+    //   (item) => Number(item.total_stake) + lovelaceDelegation < SATURATION_LIMIT,
+    // );
+    return pools;
   }
 
-  const { urlParams: { selectedPoolIds, totalAda } } = props
+  const {
+    urlParams: { selectedPoolIds, totalAda },
+  } = props;
+
   return (
     <>
       <Alert title={alertText} />
       <Header>
         <Search filter={filterSearch} />
         <ProviderSelect />
-        <SortSelect filter={filterSelect} />
-        {/* <ColorButton type="button" onClick={() => setOpenModal(true)}> */}
-        {/*  Colors meaning */}
-        {/* </ColorButton> */}
       </Header>
       <DesktopOnly>
-        <AdapoolsDesktopTable
+        <DaedalusSimpleDesktopTable
           status={status}
           delegateFunction={delegateFunction}
           data={filterPools(rowData, totalAda)}
@@ -200,7 +170,7 @@ function AdapoolsTable(props: AdapoolsTableProps): Node {
         />
       </DesktopOnly>
       <MobileOnly>
-        <AdapoolsMobileTable
+        <DaedalusSimpleMobileTable
           status={status}
           delegateFunction={delegateFunction}
           data={filterPools(rowData, totalAda)}
@@ -212,12 +182,18 @@ function AdapoolsTable(props: AdapoolsTableProps): Node {
         <Modal
           title=""
           isOpen={openModal && confirmDelegationModal}
-          onClose={() => {setOpenModal(false); setConfirmDelegationModal(false)}}
+          onClose={() => {
+            setOpenModal(false);
+            setConfirmDelegationModal(false);
+          }}
         >
           <SaturatedPoolAlert
             delegation={delegationModalData}
             onSuccess={confirmedDelegateFunction}
-            close={() => {setOpenModal(false); setConfirmDelegationModal(false)}}
+            close={() => {
+              setOpenModal(false);
+              setConfirmDelegationModal(false);
+            }}
           />
         </Modal>
       )}
@@ -225,4 +201,4 @@ function AdapoolsTable(props: AdapoolsTableProps): Node {
   );
 }
 
-export default AdapoolsTable;
+export default DaedalusSimpleContainer;
